% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{bigint}
\alias{bigint}
\title{Postgres utilities}
\usage{
bigint(x)
}
\arguments{
\item{x}{A numeric or character vector containing only whole numbers.}
}
\value{
An object of class \code{pg_bigint} that can be inserted into an
SQL query.
}
\description{
Utilities to ease the conversion from R to Postgres. Currently I can think
of only one important function:

\itemize{
 \item{\code{bigint} signals a character vector as a big integer.
 This is necessary to work with big integers in Postgres.}
}
}
\examples{
bigint(1)
bigint(100000000000000000000000) # not useful
bigint("100000000000000000000000") # better
}
