% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstore.R
\name{hstore}
\alias{hstore}
\alias{\%->\%}
\alias{\%?\%}
\alias{\%?&\%}
\alias{\%?|\%}
\alias{\%-\%}
\alias{hstore_zip}
\alias{hstore_2d}
\title{hstore operations}
\usage{
hstore \%->\% y

hstore \%?\% key

hstore \%?&\% key

hstore \%?|\% key

hstore \%-\% key

hstore_zip(hstore)

hstore_2d(hstore)
}
\arguments{
\item{hstore}{Name of a database column containing the \code{hstore} data type.}

\item{key}{Character string containing an \code{hstore} key.}
}
\value{
A character string of class \code{sql}.
}
\description{
Due to the way \code{dbplyr} translates infix operators, there is a number
of operators you can use to control Postgres/PostGIS queries. Some useful
ones include:

\itemize{
 \item{\code{\%->\%}: Extract keys \code{y} from hstore \code{x}}
 \item{\code{\%?\%}: Check if key \code{y} is in hstore \code{x}}
 \item{\code{\%?&\%}: Check if all keys \code{y} are in hstore \code{x}}
 \item{\code{\%?|\%}: Check if any keys \code{y} are in hstore \code{x}}
 \item{\code{\%-\%}: Remove a key \code{y} from hstore \code{x}}
 \item{\code{hstore_zip}: Interweave keys and values of hstore \code{x}}
 \item{\code{hstore_2d}: Convert hstore \code{x} to a 2D key-value array}
}

Note that in order to use (non-infix) functions, you need to inject them
using \code{\link[rlang:!!]{!!}}.
}
\examples{
library(dplyr)

pp_tbl("point") |>
  filter(amenity == "fast_food" & way \%&&\% !!pg_bbox(bbox)) |>
  mutate(phone = tags \%->\% "phone") |>
  select(name, way, phone)
}
