% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{list_tables}
\alias{list_tables}
\title{List available geometry tables}
\usage{
list_tables()
}
\value{
A vector.
}
\description{
Retrieve a vector of available geometry tables provided by Postpass.
Scraped from \href{https://github.com/woodpeck/postpass-ops/blob/main/SCHEMA.md}{SCHEMA.md}.
}
\examples{
list_tables()
}
