% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{local_tables}
\alias{local_tables}
\title{List available geometry tables}
\usage{
local_tables()
}
\value{
A vector.
}
\description{
Retrieve a vector of available geometry tables provided by Postpass.
Should almost always return the same tables as \code{\link{pp_tables}} but
can be run locally. Also abbreviates the table names for convenience.
}
\examples{
list_tables()
}
