% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgis.R
\name{pg_geom}
\alias{pg_geom}
\alias{pg_bbox}
\title{PostGIS utilities}
\usage{
pg_geom(x)

pg_bbox(x)
}
\arguments{
\item{x}{A \code{\link[sf:st_sfc]{sfc}} object or an object coercible to
\code{sfc}.}
}
\value{
Objects of class \code{call}.
}
\description{
\code{dbplyr} does not recognize PostGIS commands by default. These functions
fill this gap by providing simple wrappers that translate sf to PostGIS.
These functions return calls and are designed to be used inside a tidy
evaluation context.

Currently supported:

\itemize{
 \item{\code{pg_geom}: \code{st_geomfromtext} using \code{\link[sf:st_sfc]{sfc}}}
 \item{\code{pg_bbox}: \code{st_makebox2d} using \code{\link[sf]{st_bbox}}}
}

Note that to use any of these functions inside a dplyr query, you need to
defuse them using the \link[rlang:!!]{bang-bang operator} (\code{!!}). To use
PostGIS operators (e.g. \code{&&}, \code{&>} or \code{>>}), wrap them as
infix operators, i.e.

\preformatted{way \%&&\% {!!pg_bbox(geometry)}}
}
\examples{
library(sf)
nc <- system.file("shape/nc.shp", package = "sf")

pg_geom(nc[1,])
pg_bbox(nc)
}
