% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{postpass}
\alias{postpass}
\title{Send SQL to Postpass}
\usage{
postpass(
  sql,
  geojson = collection,
  collection = TRUE,
  pretty = FALSE,
  parse = TRUE,
  unwrap = TRUE,
  endpoint = "interpreter"
)
}
\arguments{
\item{sql}{Character string containing a valid Postpass SQL query.}

\item{geojson}{Whether to generate a GeoJSON or not. If \code{TRUE}, returns
an sf tibble, otherwise a normal tibble. Should only be \code{TRUE} if
\code{collection} is also \code{TRUE}, otherwise Postpass will not return
a valid GeoJSON.}

\item{collection}{Whether to aggregate results to a JSON or not.
If \code{FALSE}, a scalar must be returned by \code{x}, i.e. one row and one
column. Necessary to compute single counts.}

\item{pretty}{Whether to return a pretty unformatted JSON. Ignored if
\code{parse} is \code{TRUE}.}

\item{parse}{Whether to parse the response as a (sf) tibble. If \code{FALSE},
returns the unparsed JSON as a character string.}

\item{unwrap}{Whether to convert the \code{tags} column to a dataframe.
If \code{FALSE}, leaves tags as an unparsed JSON. If \code{TRUE}, parses it
and inserts it to the position of the \code{tags} column using
\code{\link{unwrap_tags}}.}
}
\value{
A (sf) tibble or a character string if \code{parse} is \code{FALSE}.
}
\description{
Low-level function to send an SQL query directly to Postpass.
}
\section{Limitations}{

Postpass manages API overload by regulating workers. Depending on their
estimated complexity, queries are put in a slow, medium or quick queue.

\itemize{
 \item{Quick queries: Under 100s, 10 workers}
 \item{Medium queries: Under 50,000s, 4 workers}
 \item{Slow queries: Over 50,000s, 2 workers}
}

Additionally, the PostGIS database is read-only. Queries can only retrieve
data but not write, e.g., using \code{DROP TABLE}.
}

\examples{
\donttest{sql <- "
SELECT
  name, way
FROM
  planet_osm_point
WHERE
  amenity = 'fast_food'
  AND tags -> 'addr:city' = 'Karlsruhe'
"

postpass(sql)}
}
