% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{postpass_utils}
\alias{postpass_utils}
\alias{pp_last_import}
\alias{pp_tables}
\alias{pp_schema}
\alias{pp_columns}
\alias{pp_peek}
\title{Postpass utilities}
\usage{
pp_last_import()

pp_tables()

pp_schema(table)

pp_columns(table)

pp_peek(table, limit = 10)
}
\arguments{
\item{table}{A table name. A list of names can be retrieved using
\code{pp_tables}.}

\item{limit}{Number of rows to return. Defaults to 10.}
}
\value{
\code{pp_last_import} returns a date-time object. \code{pp_tables}
and \code{pp_columns} return a character vector. \code{pp_peek} and
\code{pp_schema} return a dataframe.
}
\description{
Some possibly useful functions to navigate the PostGIS database behind
Postpass.

\itemize{
 \item{\code{pp_last_import} returns the date-time of the last OSM import.}
 \item{\code{pp_tables} returns the names of the database tables.}
 \item{\code{pp_schema} returns the schema of a given table.}
 \item{\code{pp_columns} returns the column names of a given table.}
 \item{\code{pp_peek} returns the first few rows of a given table.}
}
}
\examples{
\donttest{pp_last_import()
pp_tables()
pp_schema("planet_osm_point")
pp_columns("planet_osm_point")
pp_peek("planet_osm_point", 5)}
}
