% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{pp_tbl}
\alias{pp_tbl}
\title{Postpass tibble}
\usage{
pp_tbl(table = local_tables(), schema = NULL, name = NULL)
}
\arguments{
\item{table}{A table name (or at least its suffix). You can retrieve a list
of available table names by running \code{\link{list_tables}} or
\code{\link{pp_tables}}.}

\item{schema}{Optionally, a schema dataframe. A schema should contain only
columns accessible in the remote PostGIS database along with their prototype
classes. Schemas can be retrieved from \code{\link{pp_schema}}. If
\code{NULL}, infers a schema from \code{table}.}

\item{name}{If \code{schema} is provided, specifies the name of the Postpass
table from which to query.}
}
\value{
A lazy tibble of class \code{pp_tbl}.
}
\description{
Creates a lazy tibble based on a simulated Postgres database and a Postpass
database schema. This tibble can be manipulated using \code{dplyr} functions.
Each operation extends the SQL query which is then sent to the Postpass
server by running \code{\link{collect}}.
}
\details{
Note that all columns can be treated as either numeric or character vectors
with two exceptions:

\itemize{
 \item{\code{tags} is an
 \href{https://www.postgresql.org/docs/current/hstore.html}{\code{hstore}}
 which is something like a named list in Postgres. You can interact with it
 using \link[hstore]{hstore helpers}}.
 \item{\code{way} is a
 \href{https://postgis.net/docs/}{PostGIS} geometry. You can interact with
 it using \link[postgis]{PostGIS helpers}.}
}
}
\examples{
pp_tbl("point")
pp_tbl("roads")

# provide a custom schema
pp_tbl(schema = dplyr::tibble(amenity = character(1), way = list(NULL)))
}
