% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_by}
\alias{filter}
\alias{select}
\alias{relocate}
\alias{arrange}
\alias{distinct}
\alias{rename}
\alias{rename_with}
\alias{explain}
\alias{collect}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr:compute]{collect}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{explain}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:rename]{rename_with}}, \code{\link[dplyr]{select}}}
}}

