% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypass-package.R
\docType{package}
\name{tidypass-package}
\alias{tidypass}
\alias{tidypass-package}
\title{tidypass: Query 'OpenStreetMap' using 'dplyr'}
\description{
A tidy R interface to the 'Postpass' \url{https://postpass.geofabrik.de/} service. Allows you to use dplyr syntax to retrieve OpenStreetMap data using 'dbplyr'.
}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org}

}
\keyword{internal}
