% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{unwrap_expr}
\alias{unwrap_expr}
\title{Given a Postpass query, extracts all unknown variables (that are neither
in the postpass scheme nor in the search path) from the tags jsonb}
\usage{
unwrap_expr(.data, ...)
}
\arguments{
\item{.data}{A `pp_tbl`.}

\item{...}{Expressions.}
}
\value{
A `pp_tbl`.
}
\description{
Given a Postpass query, extracts all unknown variables (that are neither
in the postpass scheme nor in the search path) from the tags jsonb
}
