% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{unwrap_tags}
\alias{unwrap_tags}
\title{Unwrap tags}
\usage{
unwrap_tags(x, keep = FALSE)
}
\arguments{
\item{x}{A dataframe returned by \code{\link{collect}} containing a
\code{tags} column.}

\item{keep}{Whether to keep the unparsed \code{tags} column or replace it.}
}
\value{
A (sf) dataframe.
}
\description{
Helper function to turn the \code{tags} column returned by
\code{\link{collect}} to multiple columns in the same dataframe.
}
\examples{
\donttest{nc <- system.file("shape/nc.shp", package = "sf")
nc <- sf::read_sf(nc)
surry <- nc[nc$NAME \%in\% "Surry", ]

res <- pp_tbl("point") |>
  filter(amenity == "fast_food" & way \%&&\% !!pg_bbox(surry)) |>
  select(name, way, tags) |>
  collect(unwrap = FALSE)

unwrap_tags(res)}
}
