% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{var_known}
\alias{var_known}
\title{Determines whether a symbol is found in the search path of an environment}
\usage{
var_known(sym, env)
}
\description{
Determines whether a symbol is found in the search path of an environment
}
