% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{collect.pp_tbl}
\alias{collect.pp_tbl}
\title{Send SQL query to Postpass}
\usage{
\method{collect}{pp_tbl}(x, geometry = TRUE, collection = TRUE, ...)
}
\arguments{
\item{x}{A lazy tibble of class \code{\link{pp_tbl}}.}

\item{geometry}{Whether to generate a geojson or not. If \code{TRUE}, returns
an sf tibble, otherwise a normal tibble.}

\item{collection}{Whether to aggregate results to a JSON or not
\href{https://www.postgresql.org/docs/9.5/functions-aggregate.html}{\code{json_agg}}.
If \code{FALSE}, a scalar must be returned by \code{x}.}
}
\value{
A (sf) tibble.
}
\description{
Accepts a \code{dbplyr} query, sanitizes it, and sends it to Postpass.
If the query is valid, returns a (sf) tibble.
}
