% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{pp_tbl}
\alias{pp_tbl}
\title{Postpass tibble}
\usage{
pp_tbl(table = list_tables(), schema = NULL, name = NULL)
}
\arguments{
\item{table}{A table name (or at least its suffix). You can retrieve a list
of available table names by running \code{\link{list_tables}} or by taking
a look at \href{https://github.com/woodpeck/postpass-ops/blob/main/SCHEMA.md}{SCHEMA.md}.}

\item{schema}{Optionally, a schema dataframe. A schema should contain only
columns accessible in the remote PostGIS database along with their prototype
classes. If \code{NULL}, infers a schema from \code{table}.}

\item{name}{If \code{schema} is provided, specifies the name of the Postpass
table from which to query.}
}
\value{
A lazy tibble of class \code{pp_tbl}.
}
\description{
Creates a lazy tibble based on a simulated Postgres database and a Postpass
database schema. This tibble can be manipulated using \code{dplyr} functions.
Each operation extends the SQL query which is then sent to the Postpass
server by running \code{\link{collect}}. An object
}
\examples{
pp_tbl("point")
pp_tbl("roads")

# provide a custom schema
pp_tbl(schema = dplyr::tibble(amenity = character(1), way = list(NULL)))
}
