% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstore.R
\name{unwrap}
\alias{unwrap}
\title{Unwrap jsonb}
\usage{
unwrap(query, ...)
}
\arguments{
\item{query}{A lazy dataframe of class \code{pp_tbl}.}

\item{...}{One or several calls to \code{\link{\%->>\%}} where the left-hand
side can contain multiple values.}
}
\value{
Another \code{pp_tbl}.
}
\description{
\code{dbplyr} helper function to extract multiple keys from a jsonb. This
is similar to \code{unwrap_tags} but instead of unwrapping the output of
\code{\link{collect}}, it generates an SQL query to extract specific keys
inside the database. This is mostly useful to reduce the amount of typing
required. So instead of:

\preformatted{
 pp_tbl("point") |>
   transmute(
     amenity = tags %->>% "amenity",
     cuisine = tags %->>% "cuisine",
     name = tags %->>% "name",
     email = tags %->>% "email"
   )
}

you can write:

\preformatted{
 pp_tbl("point") |>
   unwrap(tags \%->>\% c("amenity", "cuisine", "name", "email"))
}
}
\examples{
pp_tbl("point") |>
   unwrap(tags \%->>\% c("amenity", "cuisine", "name", "email"))
}
